backend/ingestor/requirements.txt
paho-mqtt==1.6.1
asyncpg==0.29.0
pydantic==2.7.1
httpx==0.27.0
